# api/openapi.yaml
openapi: 3.1.0
info:
  title: KubeVirt Shepherd API
  description: |
    API for managing KubeVirt virtual machines and cloud-native governance.
    Adheres to ADR-0021: API Contract-First Development.
  version: 1.0.0
  contact:
    name: KubeVirt Shepherd Team
    url: https://github.com/kv-shepherd/shepherd
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /api/v1
    description: Production API
...
paths:
...
  /tickets:
    get:
      tags: [tickets]
      summary: List work-order tickets
      operationId: listTickets
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, APPROVED, REJECTED, CANCELLED, EXECUTING, SUCCESS, FAILED]
        - name: mine
          in: query
          description: |
            When true, return only tickets requested by the current
            authenticated user. This does not require ticket:view.
          schema:
            type: boolean
        - name: operation_type
          in: query
          description: Filter by ticket operation type.
          schema:
            type: string
            enum: [CREATE, MODIFY, DELETE, POWER, VNC_ACCESS]
        - name: selected_cluster_id
          in: query
          description: Filter by the cluster selected during CREATE execution planning.
          schema:
            type: string
        - name: placement_advisory_code
          in: query
          description: Filter by details from placement_evaluation.advisory_code persisted on CREATE tickets.
          schema:
            type: string
        - name: placement_snapshot
          in: query
          description: Filter by whether a placement evaluation snapshot exists on the ticket.
          schema:
            type: string
            enum: [present, missing]
      responses:
        '200':
          description: Ticket list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketList'

  # ── Built-In Approval Provider ─────────────────────
  /builtin-approval/tasks:
    get:
      tags: [builtin-approval]
      summary: List built-in approval tasks
      operationId: listBuiltinApprovalTasks
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, APPROVED, REJECTED, CANCELLED, EXECUTING, SUCCESS, FAILED]
        - name: operation_type
          in: query
          description: Filter by task operation type.
          schema:
            type: string
            enum: [CREATE, MODIFY, DELETE, POWER, VNC_ACCESS]
        - name: selected_cluster_id
          in: query
          description: Filter by the cluster selected during CREATE execution planning.
          schema:
            type: string
        - name: placement_advisory_code
          in: query
          description: Filter by details from placement_evaluation.advisory_code persisted on CREATE tickets.
          schema:
            type: string
        - name: placement_snapshot
          in: query
          description: Filter by whether a placement evaluation snapshot exists on the task.
          schema:
            type: string
            enum: [present, missing]
      responses:
        '200':
          description: Built-in approval task list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketList'
...
  /admin/auth-providers:
    get:
      tags: [auth-providers, admin]
      summary: List authentication providers
      operationId: listAuthProviders
      responses:
        '200':
          description: Authentication provider list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthProviderList'
    post:
      tags: [auth-providers, admin]
      summary: Create authentication provider from registered plugin type
      operationId: createAuthProvider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthProviderCreateRequest'
      responses:
        '201':
          description: Authentication provider created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthProvider'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /admin/instance-sizes:
    get:
      tags: [instance-sizes, admin]
      summary: List instance sizes for admin management
      operationId: listAdminInstanceSizes
      responses:
        '200':
          description: Instance size list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstanceSizeList'
    post:
      tags: [instance-sizes, admin]
      summary: Create instance size
      operationId: createAdminInstanceSize
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstanceSizeCreateRequest'
      responses:
        '201':
          description: Instance size created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstanceSize'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /audit-logs:
    get:
      tags: [audit, admin]
      summary: List audit logs
      operationId: listAuditLogs
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - name: action
          in: query
          schema:
            type: string
        - name: actor
          in: query
          schema:
            type: string
        - name: resource_type
          in: query
          schema:
            type: string
        - name: resource_id
          in: query
          schema:
            type: string
        - name: approval_decision
          in: query
          description: Filter approval audit entries by details.decision (for example approved, rejected, validation_failed)
          schema:
            type: string
        - name: placement_reason_code
          in: query
          description: Filter approval audit entries by details.placement_evaluation.reason_code
          schema:
            type: string
        - name: placement_advisory_code
          in: query
          description: Filter approval audit entries by details.placement_evaluation.advisory_code
          schema:
            type: string
      responses:
        '200':
          description: Audit log list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogList'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Reusable Parameters (ADR-0023 Pagination) ──────
  parameters:
    Page:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: per_page
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      required: [code]
      properties:
        code:
          type: string
          description: Machine-readable error code (frontend handles i18n)
        message:
          type: string
          description: English log message (not for UI display)
        params:
          type: object
          additionalProperties: true
        field_errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      required: [field, code]
      properties:
        field:
          type: string
        code:
          type: string
        message:
          type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer

    # ── System ──────────────────────────────────────
...
    ProvisioningCondition:
      type: object
      properties:
        type:
          type: string
        status:
          type: string
        reason:
          type: string
        message:
          type: string
        last_transition_time:
          type: string
          format: date-time
...
    ProvisioningStatus:
      type: object
      properties:
        root_data_volume_name:
          type: string
        claim_name:
          type: string
        phase:
          type: string
        progress:
          type: string
        restart_count:
          type: integer
        pvc_phase:
          type: string
        clone_type:
          type: string
          description: CDI clone execution type observed on the root PVC, for example `copy` when host-assisted fallback is used.
        clone_phase:
          type: string
          description: CDI clone phase annotation observed on the root PVC.
        clone_fallback_reason:
          type: string
          description: CDI fallback reason observed on the root PVC when efficient clone prerequisites were not met.
        failure_message:
          type: string
        conditions:
          type: array
          items:
            $ref: '#/components/schemas/ProvisioningCondition'
        recent_events:
          type: array
          items:
            $ref: '#/components/schemas/ProvisioningEvent'
...
    VMRequestPrefill:
      type: object
      required: [system_id, service_id, template_id, instance_size_id, namespace, reason, batch_count]
      properties:
        system_id:
          type: string
          format: uuid
        service_id:
          type: string
          format: uuid
        template_id:
          type: string
          format: uuid
        instance_size_id:
          type: string
          format: uuid
        namespace:
          type: string
        reason:
          type: string
        batch_count:
          type: integer
          minimum: 1
...
    TicketSummary:
      type: object
      properties:
        system_id:
          type: string
        system_name:
          type: string
        service_id:
          type: string
        service_name:
          type: string
        namespace:
          type: string
        cluster_id:
          type: string
        cluster_name:
          type: string
        cluster_environment:
          type: string
        vm_id:
          type: string
        vm_name:
          type: string
        request_vm_status:
          type: string
          description: VM status captured when the approval request was created.
        latest_vm_status:
          type: string
          description: Latest VM status resolved when the ticket was fetched.
        vm_status:
          type: string
          description: Deprecated alias for latest_vm_status.
        template_id:
          type: string
        template_name:
          type: string
        instance_size_id:
          type: string
        instance_size_name:
          type: string
        batch_count:
          type: integer
          minimum: 1
        current_cpu_cores:
          type: number
          format: double
        current_memory_gi:
          type: number
          format: double
        current_disk_gb:
          type: integer
        target_cpu_cores:
          type: number
          format: double
        target_memory_gi:
          type: number
          format: double
        target_disk_gb:
          type: integer
        power_action:
          type: string
        irreversible:
          type: boolean
        items:
          type: array
          items:
            $ref: '#/components/schemas/TicketItemSummary'
    Ticket:
      type: object
      required: [id, event_id, status, requester]
      properties:
        id:
          type: string
        event_id:
          type: string
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED, CANCELLED, EXECUTING, SUCCESS, FAILED]
        operation_type:
          type: string
          enum: [CREATE, MODIFY, DELETE, POWER, VNC_ACCESS]
          description: Type of operation this ticket represents (ADR-0015)
        requester:
          type: string
        approver:
          type: string
        reason:
          type: string
        reject_reason:
          type: string
        target_vm_id:
          type: string
          description: For DELETE tickets, the VM being deleted
        target_vm_name:
          type: string
          description: For DELETE tickets, the VM name (for display)
        summary:
          x-go-type-skip-optional-pointer: false
          allOf:
            - $ref: '#/components/schemas/TicketSummary'
        request_prefill:
          $ref: '#/components/schemas/VMRequestPrefill'
        ticket_payload:
          type: ["object", "null"]
          description: Contextual payload of the original request depending on operation_type
          additionalProperties: true
        provisioning:
          x-go-type-skip-optional-pointer: false
          allOf:
            - $ref: '#/components/schemas/ProvisioningStatus'
        placement_evaluation:
          x-go-type-skip-optional-pointer: false
          allOf:
            - $ref: '#/components/schemas/PlacementEvaluation'
        created_at:
          type: string
          format: date-time
    TicketList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Ticket'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    InstanceSize:
      type: object
      required: [id, name, cpu_cores, memory_gi]
      properties:
        id:
          type: string
        name:
          type: string
        display_name:
          type: string
        description:
          type: string
        catalog_scope:
          type: string
          enum: [unclassified, test, prod, all]
          description: Catalog visibility scope only. Not scheduling environment.
        cpu_cores:
          type: number
        cpu_request:
          description: CPU request override in cores. Omitted when no explicit override is stored.
          type: number
        memory_gi:
          type: number
        memory_request_gi:
          description: Memory request override in Gi. Omitted when no explicit override is stored.
          type: number
        disk_gb:
          type: integer
        dedicated_cpu:
          type: boolean
        requires_gpu:
          type: boolean
        requires_sriov:
          type: boolean
        requires_hugepages:
          type: boolean
        hugepages_size:
          type: string
        dv_access_modes:
          type: array
          items:
            type: string
          description: >
            Explicit CDI DataVolume PVC accessModes for the root volume. When empty together
            with `dv_volume_mode`, the platform treats the storage mode as approval-time auto resolution.
        dv_volume_mode:
          type: string
          enum: [Block, Filesystem]
          description: >
            Explicit CDI DataVolume PVC volumeMode for the root volume. When empty together
            with `dv_access_modes`, the platform treats the storage mode as approval-time auto resolution.
        spec_overrides:
          type: object
          additionalProperties: true
          description: >
            KubeVirt spec path overrides (admin-only). Only populated in admin
            catalog endpoints; omitted from user-facing responses.
        sort_order:
          type: integer
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
    InstanceSizeCreateRequest:
      type: object
      required: [name, cpu_cores, memory_gi]
      properties:
        name:
          x-oapi-codegen-extra-tags:
            validate: "required"
          type: string
        display_name:
          type: string
        description:
          type: string
        catalog_scope:
          type: string
          enum: [unclassified, test, prod, all]
          description: Catalog visibility scope only. Not scheduling environment.
        cpu_cores:
          x-oapi-codegen-extra-tags:
            validate: "required,gte=0.5"
          type: number
          minimum: 0.5
          multipleOf: 0.5
        memory_gi:
          x-oapi-codegen-extra-tags:
            validate: "required,gte=0.5"
          type: number
          minimum: 0.5
          multipleOf: 0.5
        disk_gb:
          x-oapi-codegen-extra-tags:
            validate: "gte=1"
          type: integer
          minimum: 1
        cpu_request:
          x-oapi-codegen-extra-tags:
            validate: "gte=0.5"
          type: number
          minimum: 0.5
          multipleOf: 0.5
        memory_request_gi:
          x-oapi-codegen-extra-tags:
            validate: "gte=0.5"
          type: number
          minimum: 0.5
          multipleOf: 0.5
        dedicated_cpu:
          type: boolean
        requires_gpu:
          type: boolean
        requires_sriov:
          type: boolean
        requires_hugepages:
          type: boolean
        hugepages_size:
          type: string
        dv_access_modes:
          type: array
          items:
            type: string
          description: >
            Explicit CDI DataVolume PVC accessModes for the root volume. On update,
            send an empty array to clear an existing explicit root volume mode and
            return the size to approval-time auto resolution.
        dv_volume_mode:
          type: string
          enum: [Block, Filesystem]
        spec_overrides:
          type: object
          additionalProperties: true
        sort_order:
          type: integer
        enabled:
          type: boolean
...
    InstanceSizeList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/InstanceSize'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    AuthProvider:
      type: object
      required: [id, name, auth_type, enabled]
      properties:
        id:
          type: string
        name:
          type: string
        auth_type:
          type: string
          description: Registered auth provider plugin type key
        config:
          type: object
          additionalProperties: true
        enabled:
          type: boolean
        sort_order:
          type: integer
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AuthProviderList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AuthProvider'
        pagination:
          $ref: '#/components/schemas/Pagination'
    AuthProviderCreateRequest:
      type: object
      required: [name, auth_type]
      properties:
        name:
          x-oapi-codegen-extra-tags:
            validate: "required"
          type: string
        auth_type:
          x-oapi-codegen-extra-tags:
            validate: "required"
          type: string
          description: Registered auth provider plugin type key
        config:
          type: object
          additionalProperties: true
        enabled:
          type: boolean
        sort_order:
          type: integer
...
    AuditLog:
      type: object
      required: [id, action, resource_type, resource_id, actor, created_at]
      properties:
        id:
          type: string
        action:
          type: string
        resource_type:
          type: string
        resource_id:
          type: string
        actor:
          type: string
        approval_decision:
          type: string
        placement_summary:
          x-go-type-skip-optional-pointer: false
          allOf:
            - $ref: '#/components/schemas/AuditPlacementSummary'
        details:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
    AuditPlacementSummary:
      type: object
      properties:
        selected_cluster_name:
          type: string
        selected_cluster_id:
          type: string
        eligible:
          x-go-type-skip-optional-pointer: false
          type: ["boolean", "null"]
        reason_code:
          type: string
        advisory_code:
          type: string
    AuditLogList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AuditLog'
        pagination:
          $ref: '#/components/schemas/Pagination'
    # ── Namespace Registry (ADR-0017) ────────────────
...
security:
  - BearerAuth: []